moved firebase config to separate file
