+++ Lost-Seeker/src/config/firebase.js
@@ -1,0 +1,13 @@
+import { initializeApp } from "firebase/app"
+import { getAuth } from "firebase/auth"
+const firebaseConfig = {
+  apiKey: "AIzaSyB2DxDn0VT7kCGQZKmeXmKmG5zWcM8qHSA",
+  authDomain: "lost-seeker.firebaseapp.com",
+  projectId: "lost-seeker",
+  storageBucket: "lost-seeker.appspot.com",
+  messagingSenderId: "727937133568",
+  appId: "1:727937133568:web:12028899a4907ffa10fce5",
+  measurementId: "G-Z3J14HNCXK",
+}
+const app = initializeApp(firebaseConfig)
+export const auth = getAuth(app)
